[FABG-988] Rename cauthdsl to policydsl
This change renames the cauthdsl to policydsl to match with Fabric.
