+++ pkg/client/common/selection/dynamicselection/pgresolver/pgresolver_test.go
@@ -10,7 +10,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/hyperledger/fabric-sdk-go/internal/github.com/hyperledger/fabric/common/cauthdsl"
+	"github.com/hyperledger/fabric-sdk-go/internal/github.com/hyperledger/fabric/common/policydsl"
 	"github.com/hyperledger/fabric-sdk-go/pkg/common/providers/fab"
 	mocks "github.com/hyperledger/fabric-sdk-go/pkg/fab/mocks"
 	common "github.com/hyperledger/fabric-protos-go/common"
@@ -103,7 +103,7 @@ func TestPeerGroupResolverPolicy1(t *testing.T) {
 
 func TestPeerGroupResolverAcceptAllPolicy(t *testing.T) {
 
-	sigPolicyEnv := cauthdsl.AcceptAllPolicy
+	sigPolicyEnv := policydsl.AcceptAllPolicy
 
 	expected := []PeerGroup{
 		pg(p1), pg(p2), pg(p3), pg(p4), pg(p5), pg(p6),
